requirements.txt
flask_sqlalchemy
flask_mail
werkzeug
cryptography
flask_sqlalchemy
flask_mail
werkzeug
cryptography
